+++ server/src/__tests__/semanticTokens.test.ts
@@ -5,7 +5,7 @@
 
 import { analyzer } from '../tree-sitter/analyzer'
 import { generateParser } from '../tree-sitter/parser'
-import { getParsedTokens, TOKEN_LEGEND } from '../semanticTokensProvider'
+import { getParsedTokens, TOKEN_LEGEND } from '../semanticTokens'
 import { FIXTURE_DOCUMENT } from './fixtures/fixtures'
 
 const DUMMY_URI = 'dummy_uri'
